New flag in config.php $dPconfig["restrict_task_time_editing"] = false;
Now the sysadmin is able to configure if a normal user is able to change time information from tasks. If this flag is set up to true, the user needs to be the task owner, project owner or sys admin.

Excellent for controlling real start and due dates! no more "it was due today!! look at dotproject's information! I swear I didn't change it!"