Added a check for missing modules.
If an external module is installed and then its folder is deleted, but its entries are left in the database, the module is still showing in the menu. If users clicked on that, it was showing a PHP error message. Now it is showing a Warning message, suggesting that the according module is probably missing.

I wasn't sure if there was a proper warning page (maybe one as part of the public module) so I quickly made one up. If there is any standard for that, this should be changed to comply with it.