WARNING: From this change on, CVS is to be considered unstable ... sorry guys but this has to be done, and sooner rather than later.
Now for the change - smarty going in.
dp functions to handle html, wrapped up in smarty functions (later to be replaced with pure smarty functions).
Converted functions so far - arraySelect(), dPformSafe() and $AppUI->_()

The only file converted in smarty so far is modules/companies/addedit.php (and the resulting style/smarty1/companies/addedit.html template) but for that change, a lot of the ground work (I believe) has been made. To help with smartification, use the $smarty variable already declared, and put theme files in directories, named after the module/file they relate to in smarty1.