Fixes for Mantis Reports #2486 and #2545:
SMTPSend() in classes/libmail.class.php now tries to connect first using the
ESMTP keyword "EHLO" and then falls back to older SMTP "HELO".

Also, since ESMTP functions are potentially available, capability to apply TLS
to a connection has been added should it be desired or required. Use of TLS is
toggled by a new config value. If TLS is selected, the fallback to "HELO" is
not done as the keyword "STARTTLS" is only available under ESMTP.

git-svn-id: https://dotproject.svn.sourceforge.net/svnroot/dotproject/branches/stable_2@5880 6cfb33f2-fe3e-0410-9678-87ca8a581921